+++ src/components/survey/SurveySingleItemView/stories/SurveySingleItem.stories.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import SurveySingleItemView from "../SurveySingleItemView";
 
 import 'localStyles';
+import { CommonResponseComponentProps } from "../utils";
 
 export default {
   title: "Survey Item Types/Item Properties"
@@ -225,6 +226,69 @@ export const Example = () =>
                 key: 'rg', role: 'responseGroup', items: [
                   {
                     key: 'scg', role: 'singleChoiceGroup',
+                    items: [
+                      { key: '1', role: 'option', content: [{ code: 'en', resolvedText: 'Option 1' }] },
+                      { key: '1b', role: 'option', content: [{ code: 'en', resolvedText: 'Option 1' }] },
+                      { key: '2', role: 'option', content: [{ code: 'en', resolvedText: 'Option 2' }] },
+                      { key: '3', role: 'input', content: [{ code: 'en', resolvedText: 'Option 2 here' }] },
+                      { key: 't1', role: 'dateInput', content: [{ code: 'en', resolvedText: 'Subtitle' }] },
+                    ]
+                  }
+                ]
+              }
+            ]
+          },
+        }}
+        responsePrefill={undefined}
+        responseChanged={(response) => console.log(response)}
+        showInvalid={false}
+        languageCode="en"
+        invalidWarning={invalidWarning}
+        showKeys={false}
+      />
+
+
+    </div>
+
+
+    <div className={itemContainerClass}>
+      <SurveySingleItemView
+        customResponseComponents={[
+          {
+            name: 'custom',
+            component: (props: CommonResponseComponentProps) => {
+              return <div>
+                <p>Custom response component is used here</p>
+                <button className="btn btn-primary" onClick={() => props.responseChanged({
+                  key: 'custom',
+                  value: 'customValue'
+                })}>Set response</button>
+              </div>
+            }
+          }
+        ]}
+        renderItem={{
+          key: 'test.q2withhelp',
+          version: 1,
+          components: {
+            key: 'root',
+            role: 'root',
+            items: [
+              {
+                key: 't', role: 'title',
+                content: [{ code: 'en', resolvedText: "With custom response component -> " }],
+                description: [{ code: 'en', resolvedText: "Question title's description attribute" }]
+              },
+              {
+                key: 'h', role: 'helpGroup',
+                items: [
+                  { key: '1', role: 'text', content: [{ code: 'en', resolvedText: 'Some text' }] },
+                ]
+              },
+              {
+                key: 'rg', role: 'responseGroup', items: [
+                  {
+                    key: 'scg', role: 'custom',
                     items: [
                       { key: '1', role: 'option', content: [{ code: 'en', resolvedText: 'Option 1' }] },
                       { key: '1b', role: 'option', content: [{ code: 'en', resolvedText: 'Option 1' }] },