+++ src/components/survey/SurveyView/SurveyView.tsx
@@ -1,6 +1,7 @@
 import React, { useState, Fragment, useEffect } from 'react';
 import { Survey, SurveySingleItemResponse, SurveyContext } from 'survey-engine/lib/data_types';
 import { SurveyEngineCore } from 'survey-engine/lib/engine';
+import { CustomSurveyResponseComponent } from '../SurveySingleItemView/ResponseComponent/ResponseComponent';
 import SurveyPageView from './SurveyPageView/SurveyPageView';
 import SurveyProgress from './SurveyProgress/SurveyProgress';
 
@@ -18,6 +19,7 @@ interface SurveyViewProps {
   invalidResponseText: string;
   hideBackButton?: boolean;
   showKeys?: boolean;
+  customResponseComponents?: Array<CustomSurveyResponseComponent>;
   // init with temporary loaded results
   // save temporary result
 }
@@ -89,6 +91,7 @@ const SurveyView: React.FC<SurveyViewProps> = (props) => {
         onResponsesChanged();
       }}
       showKeys={props.showKeys}
+      customResponseComponents={props.customResponseComponents}
     />;
   }